+++ TACCtests/HT1.test
@@ -1,0 +1,19 @@
+0
+1
+0
+2000
+1
+1
+0
+1
+1e-05
+0
+0
+0
+1
+1
+0
+1
+0
+0
+0
