+++ TACCtests/HT3.test
@@ -1,0 +1,51 @@
+0
+1
+0
+2000
+1
+3
+0
+1
+1e-05
+0
+0
+0
+3
+3
+0
+0
+1
+0
+0
+0
+1
+3
+0
+0
+1
+1
+0
+1
+2
+1
+1
+0
+1
+1
+1
+2
+3
+3
+0
+0
+1
+1
+0
+1
+2
+1
+1
+0
+1
+1
+1
