+++ TACCtests/HT4.test
@@ -1,0 +1,79 @@
+0
+1
+0
+2000
+1
+4
+0
+1
+1e-05
+0
+0
+0
+4
+4
+1
+0
+0
+1
+1
+0
+1
+0
+1
+0
+1
+0
+0
+0
+0
+1
+0
+1
+3
+1
+1
+0
+1
+1
+0
+0
+1
+1
+0
+0
+0
+0
+1
+1
+1
+0
+3
+4
+1
+0
+0
+1
+1
+0
+1
+0
+1
+3
+1
+1
+0
+1
+1
+0
+0
+1
+1
+0
+0
+0
+0
+1
+1
+1
+0
