+++ TACCtests/HT5.test
@@ -1,0 +1,115 @@
+0
+1
+0
+2000
+1
+5
+0
+1
+1e-05
+0
+0
+0
+5
+5
+2
+0
+1
+0
+1
+0
+1
+0
+0
+0
+1
+3
+0
+0
+0
+0
+0
+0
+1
+1
+0
+1
+1
+1
+1
+1
+0
+0
+1
+0
+0
+0
+1
+3
+0
+0
+1
+1
+1
+0
+0
+0
+0
+1
+0
+0
+0
+0
+0
+1
+1
+0
+1
+1
+1
+4
+5
+3
+0
+0
+0
+0
+0
+0
+1
+1
+0
+1
+1
+1
+1
+1
+0
+0
+1
+0
+0
+0
+1
+3
+0
+0
+1
+1
+1
+0
+0
+0
+0
+1
+0
+0
+0
+0
+0
+1
+1
+0
+1
+1
+1
