+++ TACCtests/HT6.test
@@ -1,0 +1,159 @@
+0
+1
+0
+2000
+1
+6
+0
+1
+1e-05
+0
+0
+0
+6
+6
+2
+0
+1
+0
+1
+0
+1
+0
+0
+0
+1
+0
+1
+1
+0
+0
+0
+0
+0
+0
+1
+1
+0
+1
+0
+1
+1
+1
+1
+1
+0
+0
+1
+0
+0
+0
+1
+0
+1
+3
+0
+0
+1
+1
+1
+0
+0
+0
+0
+1
+0
+1
+3
+0
+0
+0
+0
+1
+1
+0
+1
+1
+0
+0
+1
+1
+0
+0
+0
+0
+0
+0
+0
+0
+1
+1
+1
+0
+5
+6
+1
+0
+0
+0
+0
+0
+0
+1
+1
+0
+1
+0
+1
+1
+1
+1
+1
+0
+0
+1
+0
+0
+0
+1
+0
+1
+3
+0
+0
+1
+1
+1
+0
+0
+0
+0
+1
+0
+1
+3
+0
+0
+0
+0
+1
+1
+0
+1
+1
+0
+0
+1
+1
+0
+0
+0
+0
+0
+0
+0
+0
+1
+1
+1
+0
