+++ TACCtests/HT7.test
@@ -1,40 +1,23 @@
 0
 1
 0
-10000
+2000
 1
-9
+7
 0
 1
 1e-05
 0
 0
 0
-9
-9
-0
+7
+7
+1
+1
+1
 0
 1
 0
-0
-0
-1
-0
-0
-0
-1
-0
-0
-0
-1
-0
-0
-0
-1
-3
-0
-0
-1
 1
 0
 1
@@ -42,31 +25,6 @@
 0
 0
 1
-0
-0
-0
-1
-0
-0
-0
-1
-3
-0
-0
-0
-0
-0
-0
-1
-1
-0
-1
-0
-0
-0
-1
-0
-0
 0
 1
 3
@@ -78,36 +36,11 @@
 0
 0
 0
-0
-0
 1
 1
 0
 1
 0
-0
-0
-1
-1
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-1
-1
-0
-1
-2
 1
 1
 0
@@ -124,27 +57,36 @@
 0
 0
 0
-0
-0
-2
-0
-0
-0
-0
 1
-1
+0
 0
 1
 1
 1
 0
 0
+1
+0
+0
+0
+1
+0
+1
+3
 0
 0
 0
 0
+1
+1
+1
 0
 0
+0
+0
+1
+0
+1
 2
 0
 0
@@ -152,8 +94,6 @@
 0
 0
 0
-0
-0
 1
 1
 0
@@ -162,11 +102,7 @@
 1
 0
 0
-0
-0
-3
-0
-0
+1
 0
 0
 0
@@ -179,39 +115,16 @@
 0
 1
 1
-0
-1
 1
 0
-8
-9
-3
-0
-0
+6
+7
+1
 1
 1
 0
 1
 0
-0
-0
-1
-0
-0
-0
-1
-0
-0
-0
-1
-3
-0
-0
-0
-0
-0
-0
-1
 1
 0
 1
@@ -219,8 +132,6 @@
 0
 0
 1
-0
-0
 0
 1
 3
@@ -232,38 +143,31 @@
 0
 0
 0
-0
-0
 1
 1
 0
 1
 0
-0
-0
 1
 1
 0
 0
+1
+1
+1
 0
 0
-0
-0
-0
-0
-0
-0
-0
+1
 0
 0
 0
 1
-1
 0
 1
-2
-1
-1
+3
+0
+0
+0
 0
 1
 1
@@ -272,33 +176,9 @@
 0
 0
 0
-0
-0
-0
-0
-0
-0
-0
-0
-2
-0
-0
-0
-0
-1
 1
 0
 1
-1
-1
-0
-0
-0
-0
-0
-0
-0
-0
 2
 0
 0
@@ -306,8 +186,6 @@
 0
 0
 0
-0
-0
 1
 1
 0
@@ -316,11 +194,7 @@
 1
 0
 0
-0
-0
-3
-0
-0
+1
 0
 0
 0
@@ -332,8 +206,6 @@
 0
 0
 1
-1
-0
 1
 1
 0
