+++ TACCtests/HT8.test
@@ -1,217 +1,21 @@
 0
 1
 0
-10000
+2000
 1
-9
+8
 0
 1
 1e-05
 0
 0
 0
-9
-9
-0
-0
-1
-0
-0
-0
-1
-0
-0
-0
-1
-0
-0
-0
-1
-0
-0
-0
-1
-3
-0
-0
-1
-1
-0
-1
-0
-0
-0
-1
-0
-0
-0
-1
-0
-0
-0
-1
-3
-0
-0
-0
-0
-0
-0
-1
-1
-0
-1
-0
-0
-0
-1
-0
-0
-0
-1
-3
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-1
-1
-0
-1
-0
-0
-0
-1
-3
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-1
-1
-0
-1
-2
-1
-1
-0
-1
-1
-1
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-2
-0
-0
-0
-0
-1
-1
-0
-1
-1
-1
-0
-0
-0
-0
-0
-0
-0
-0
-2
-0
-0
-0
-0
-0
-0
-0
-0
-1
-1
-0
-1
-1
-1
-0
-0
-0
-0
-2
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-1
-1
-0
-1
-1
-1
 8
-9
-3
-0
-0
+8
+1
 1
 1
 0
-1
-0
-0
-0
-1
-0
-0
-0
-1
-0
-0
-0
-1
-3
-0
-0
-0
-0
-0
-0
-1
 1
 0
 1
@@ -223,11 +27,9 @@
 0
 0
 1
-3
 0
-0
-0
-0
+1
+1
 0
 0
 0
@@ -242,11 +44,9 @@
 0
 0
 1
-3
 0
-0
-0
-0
+1
+1
 0
 0
 0
@@ -258,35 +58,30 @@
 0
 0
 1
+1
+0
 1
 0
 1
 2
-1
-1
 0
 1
 1
+0
+0
 1
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-2
-0
 0
 0
 0
 1
+0
+0
+0
 1
+0
+1
+1
+0
 0
 1
 1
@@ -295,15 +90,14 @@
 0
 0
 0
+1
 0
 0
 0
+1
 0
-2
-0
-0
-0
-0
+1
+1
 0
 0
 0
@@ -313,16 +107,14 @@
 0
 1
 1
+0
+0
+0
+0
 1
 0
-0
-0
-0
-2
-0
-0
-0
-0
+1
+1
 0
 0
 0
@@ -336,4 +128,144 @@
 0
 1
 1
+0
+0
 1
+1
+0
+0
+0
+0
+0
+0
+0
+0
+0
+0
+0
+0
+1
+1
+1
+0
+7
+8
+1
+1
+1
+0
+1
+0
+1
+0
+0
+0
+1
+0
+0
+0
+1
+0
+1
+1
+0
+0
+0
+0
+0
+0
+1
+1
+0
+1
+0
+0
+0
+1
+0
+1
+1
+0
+0
+0
+0
+0
+0
+0
+0
+0
+0
+1
+1
+0
+1
+0
+1
+1
+0
+0
+1
+1
+1
+0
+0
+0
+0
+1
+0
+0
+0
+1
+0
+1
+1
+0
+0
+0
+0
+1
+1
+0
+1
+1
+0
+0
+0
+0
+1
+0
+1
+1
+0
+0
+0
+0
+0
+0
+0
+0
+1
+1
+0
+1
+1
+0
+0
+1
+1
+0
+0
+0
+0
+0
+0
+0
+0
+0
+0
+0
+0
+1
+1
+1
+0
