--- TACCtests/HT9-000000001.test
@@ -1,339 +1,0 @@
-0
-1
-0
-10000
-1
-9
-0
-1
-1e-05
-0
-0
-0
-9
-9
-2
-1
-0
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-3
-1
-1
-1
-0
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-1
-0
-0
-1
-1
-1
-0
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-3
-0
-0
-0
-0
-1
-1
-1
-0
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-1
-0
-0
-0
-0
-0
-0
-1
-1
-1
-0
-0
-1
-0
-1
-0
-1
-0
-1
-3
-0
-0
-0
-0
-0
-0
-0
-0
-1
-1
-1
-0
-0
-1
-0
-1
-0
-1
-1
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-1
-1
-1
-0
-0
-1
-0
-1
-3
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-1
-1
-1
-0
-0
-1
-3
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-1
-1
-1
-0
-8
-9
-3
-1
-1
-1
-0
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-1
-0
-0
-1
-1
-1
-0
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-3
-0
-0
-0
-0
-1
-1
-1
-0
-0
-1
-0
-1
-0
-1
-0
-1
-0
-1
-1
-0
-0
-0
-0
-0
-0
-1
-1
-1
-0
-0
-1
-0
-1
-0
-1
-0
-1
-3
-0
-0
-0
-0
-0
-0
-0
-0
-1
-1
-1
-0
-0
-1
-0
-1
-0
-1
-1
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-1
-1
-1
-0
-0
-1
-0
-1
-3
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-1
-1
-1
-0
-0
-1
-3
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-1
-1
-1
-0
