+++ TACCtests/HT9.test
@@ -1,7 +1,7 @@
 0
 1
 0
-10000
+2000
 1
 9
 0
@@ -12,7 +12,30 @@
 0
 9
 9
-3
+0
+0
+1
+0
+1
+0
+0
+0
+1
+0
+0
+0
+1
+0
+0
+0
+1
+0
+1
+1
+0
+0
+0
+0
 1
 1
 0
@@ -26,31 +49,8 @@
 0
 1
 0
-0
-0
-1
-0
-1
-3
-0
-0
-0
-0
 1
 1
-0
-1
-0
-0
-0
-1
-0
-0
-0
-1
-0
-1
-3
 0
 0
 0
@@ -65,68 +65,17 @@
 1
 0
 0
-0
-1
-0
-1
-3
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-1
-1
 0
 1
 0
 1
 3
 0
-1
-1
-1
 0
 0
 0
 0
 0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-2
-0
-0
-1
-1
-0
-1
-1
-1
-0
-0
-0
-0
-0
-0
-0
-0
-0
-0
-2
 0
 0
 0
@@ -137,14 +86,65 @@
 1
 0
 1
+0
+1
+1
+1
+1
+1
+0
+0
+0
+0
+1
+0
+0
+0
+1
+0
+0
+0
+1
+0
+1
+1
+0
+0
+1
+1
+0
 1
 1
 0
 0
 0
 0
+1
 0
 0
+0
+1
+0
+1
+3
+0
+0
+0
+0
+0
+0
+1
+1
+0
+1
+1
+0
+0
+0
+0
+1
+0
+1
 0
 0
 0
@@ -185,26 +185,7 @@
 0
 8
 9
-3
 1
-1
-0
-1
-0
-0
-0
-1
-0
-0
-0
-1
-0
-0
-0
-1
-0
-1
-3
 0
 0
 0
@@ -223,7 +204,7 @@
 1
 0
 1
-3
+1
 0
 0
 0
@@ -261,12 +242,7 @@
 1
 0
 1
-2
-0
-0
 1
-1
-0
 1
 1
 1
@@ -274,13 +250,37 @@
 0
 0
 0
+1
+0
+0
+0
+1
+0
+0
+0
+1
+0
+1
+1
+0
+0
+1
+1
+0
+1
+1
 0
 0
 0
 0
+1
 0
 0
-2
+0
+1
+0
+1
+3
 0
 0
 0
@@ -292,13 +292,13 @@
 0
 1
 1
+0
+0
+0
+0
 1
 0
-0
-0
-0
-0
-0
+1
 0
 0
 0
