+++ circuits/HTstack.circ
@@ -8,6 +8,7 @@ T
 H // 3rd T: 0.5732
 T
 H // 4th T: 0.9786
+/*
 T
 H // 5th T: 0.4634
 T
@@ -18,5 +19,4 @@ T
 H // 8th T: 0.6714
 T
 H // 9th T: 0.9176
-/*
 */
