+++ libcirc/stabilizer/stabilizer.h
@@ -1,35 +1,40 @@
-#include "../utils/matrix.h"
+#include <gsl/gsl_vector.h>
+#include <gsl/gsl_matrix.h>
+#include <gsl/gsl_complex.h>
+#include <gsl/gsl_complex_math.h>
 
+int mod(int a, int b);
 
 //define K from RandomStabilizerState algorithm (page 16)
 struct StabilizerState {
 	int n;
 	int k;
-    struct BitVector* h;		//in \mathbb{F}^n_2
-	struct BitMatrix* G;		//in \mathbb{F}^{n\times n}_2
-	struct BitMatrix* Gbar;	//= (G^-1)^T
+	gsl_vector *h;		//in \mathbb{F}^n_2
+	gsl_matrix *G;		//in \mathbb{F}^{n\times n}_2
+	gsl_matrix *Gbar;	//= (G^-1)^T
 
 	int Q;				//in \mathbb{Z}_8
-	struct BitMatrix *D2  // in {0,2,4,6}^k
-	struct BitMatrix *D1; // D[i] = 2*D2[i] + D1[i]
-	struct BitMatrix* J; //in {0,4}^{k\times k}, symmetric
+	gsl_vector *D;		//in {0,2,4,6}^k
+	gsl_matrix *J;		//in {0,4}^{k\times k}, symmetric
 };
 
 
 struct StabilizerState* allocStabilizerState(int n, int k);
+
 void deepCopyState(struct StabilizerState *dest, struct StabilizerState *src);
-void freeStabilizerState(struct StabilizerState* state);
-void printStabilizerState(struct StabilizerState* state);  // prints python code for easy debugging
 
-void exponentialSumExact(struct StabilizerState* state, int* eps, int* p, int* m);
-Complex exponentialSum(struct StabilizerState* state);
+void exponentialSum(struct StabilizerState *state, int *eps, int *p, int *m, gsl_complex *ans, int exact);
 
-int shrink(struct StabilizerState* state, gsl_vector* xi, int alpha, int lazy);
+int shrink(struct StabilizerState *state, gsl_vector *xi, int alpha, int lazy);
 
-void innerProductExact(struct StabilizerState* state1, struct StabilizerState* state2, int* eps, int* p, int* m);
-Complex innerProduct(struct StabilizerState* state1, struct StabilizerState* state2);
+void innerProduct(struct StabilizerState *state1, struct StabilizerState *state2, int *eps, int *p, int *m, gsl_complex *ans, int exact);
 
 struct StabilizerState* randomStabilizerState(int n);
 
-double measurePauli(struct StabilizerState* state, int m, struct BitVector* zeta, struct BitVector* xi);
-void extend(struct StabilizerState* state, struct BitVector* xi);
+void extend(struct StabilizerState *state, gsl_vector *xi);
+
+double measurePauli(struct StabilizerState *state, int m, gsl_vector *zeta, gsl_vector *xi);
+
+void freeStabilizerState(struct StabilizerState *state);
+
+void printStabilizerState(struct StabilizerState *state);
