--- performance.txt
@@ -1,22 +1,0 @@
-{"k": 1, "t": 2, "mode": "exactDecomp", "L": 1000, "mutime": 0.31794357333333334, "sigmatime": 0.013062176495947712, "muout": 0.7510070041351488, "sigmaout": 0.0255728210359326, "ntests": 300}
-{"k": 2, "t": 4, "mode": "exactDecomp", "L": 1000, "mutime": 0.3387080866666667, "sigmatime": 0.017720938675264607, "muout": 0.9788389724042627, "sigmaout": 0.041930445779502556, "ntests": 300}
-{"k": 3, "t": 6, "mode": "exactDecomp", "L": 1000, "mutime": 0.3793576066666667, "sigmatime": 0.01586756667541127, "muout": 0.9410189875106915, "sigmaout": 0.03988578784616409, "ntests": 300}
-{"k": 4, "t": 8, "mode": "exactDecomp", "L": 1000, "mutime": 0.5024812399999999, "sigmatime": 0.024262908800383627, "muout": 0.6704154188332577, "sigmaout": 0.031195485367652216, "ntests": 300}
-{"k": 5, "t": 10, "mode": "exactDecomp", "L": 1000, "mutime": 0.9216132366666666, "sigmatime": 0.04775595888878499, "muout": 0.4694043634266732, "sigmaout": 0.019570329359202877, "ntests": 300}
-{"k": 6, "t": 12, "mode": "exactDecomp", "L": 1000, "mutime": 2.2975540899999993, "sigmatime": 0.07273659974782458, "muout": 0.5510955478115865, "sigmaout": 0.024160094987694187, "ntests": 300}
-{"k": 7, "t": 14, "mode": "exactDecomp", "L": 1000, "mutime": 6.782281553333333, "sigmatime": 0.23831756144966282, "muout": 0.8274004302336859, "sigmaout": 0.03582811519181259, "ntests": 300}
-{"k": 8, "t": 16, "mode": "exactDecomp", "L": 1000, "mutime": 20.528028990000003, "sigmatime": 0.7492651457660922, "muout": 0.9964265604478235, "sigmaout": 0.04342418828130821, "ntests": 300}
-{"k": 9, "t": 18, "mode": "exactDecomp", "L": 1000, "mutime": 60.92814355666667, "sigmatime": 2.190834933421151, "muout": 0.8778209801124786, "sigmaout": 0.040037586179292706, "ntests": 300}
-{"k": 2, "t": 2, "mode": "noapprox", "L": 1000, "mutime": 0.3148821433333333, "sigmatime": 0.013562028447204676, "muout": 0.75, "sigmaout": 0.0, "ntests": 300}
-{"k": 4, "t": 4, "mode": "noapprox", "L": 1000, "mutime": 0.3143012499999999, "sigmatime": 0.012305322681432074, "muout": 0.9785539999999999, "sigmaout": 1.1102230246251565e-16, "ntests": 300}
-{"k": 6, "t": 6, "mode": "noapprox", "L": 1000, "mutime": 0.3186680933333333, "sigmatime": 0.01577872508531521, "muout": 0.9375, "sigmaout": 0.0, "ntests": 300}
-{"k": 8, "t": 8, "mode": "noapprox", "L": 1000, "mutime": 0.3253823566666667, "sigmatime": 0.013665921239935572, "muout": 0.6714159999999999, "sigmaout": 1.1102230246251565e-16, "ntests": 300}
-{"k": 1, "t": 2, "mode": "Lsample", "L": 1000, "mutime": 0.32248950333333337, "sigmatime": 0.014818559656165043, "muout": 0.7084126975822068, "sigmaout": 0.1536929930009748, "ntests": 300}
-{"k": 2, "t": 4, "mode": "Lsample", "L": 1000, "mutime": 0.3375749433333333, "sigmatime": 0.014240151362027569, "muout": 0.8800387217903991, "sigmaout": 0.10911581375500183, "ntests": 300}
-{"k": 3, "t": 6, "mode": "Lsample", "L": 1000, "mutime": 0.3921286333333333, "sigmatime": 0.031212506162202926, "muout": 0.8450039475464374, "sigmaout": 0.11756594347710107, "ntests": 300}
-{"k": 4, "t": 8, "mode": "Lsample", "L": 1000, "mutime": 0.5234754466666668, "sigmatime": 0.044367832463476906, "muout": 0.6553770599665214, "sigmaout": 0.10864332722786606, "ntests": 300}
-{"k": 5, "t": 10, "mode": "Lsample", "L": 1000, "mutime": 0.94127477, "sigmatime": 0.08124446850912641, "muout": 0.46592921634261075, "sigmaout": 0.07670361320988647, "ntests": 300}
-{"k": 6, "t": 12, "mode": "Lsample", "L": 1000, "mutime": 2.2614273033333334, "sigmatime": 0.07494727813081664, "muout": 0.5547103796097143, "sigmaout": 0.09413966096015343, "ntests": 300}
-{"k": 7, "t": 14, "mode": "Lsample", "L": 1000, "mutime": 6.52005529, "sigmatime": 0.21528072978158674, "muout": 0.8170962458525732, "sigmaout": 0.06364730358614405, "ntests": 300}
-{"k": 8, "t": 16, "mode": "Lsample", "L": 1000, "mutime": 19.789358149999998, "sigmatime": 0.6851439596141488, "muout": 0.9784077856029988, "sigmaout": 0.05000976182398213, "ntests": 300}
-{"k": 9, "t": 18, "mode": "Lsample", "L": 1000, "mutime": 59.03102449, "sigmatime": 1.850470869921575, "muout": 0.8809963422167542, "sigmaout": 0.05081486422564056, "ntests": 100}
